Adding a example to the readme.md
I struggled with using it at first, as I thought I had to include the brackets to make it work. I think adding an example would help prevent this issues from happening to other people.  

Also, the indent in the code blocks gets interpreted, so I removed that as well for a cleaner look.